requirements.txt
customtkinter
psutil
pandas
pyobjc-framework-Quartz
pyobjc-framework-ApplicationServices
